rdbms: make table name parameter comments more complete and uniform
Better emphasize that there should only be one string that callers use
to reference a given table. This is what ChangedTablesTracker relies on
for write queries. For read queries, databases can be vertically sharded
on different servers or remapped using virtual domains, so the use of
"db.table" notation should be discouraged there as well.

Mention that "information_schema." is allowed as well as unqualified
table names.

Change-Id: I0b6295d0dd58aafd09a5c99aaf0ae9a81641d55a