installer: Add installPreConfigured.php
Add a CLI installer variant which uses the existing LocalSettings.php
for its configuration source. It runs all installer steps except for
mailing list subscription and sysop creation.

I used the name installPreConfigured.php instead of addWiki.php to make
it easier for users to discover. But AddWiki is used internally as a
prefix, because it's shorter.

* Broaden the type of ConnectionStatus::getDB() so that it can return a
  DBConnRef from the unmodified LBFactory.
* Split Installer::resetMediaWikiServices(), providing a public static
  method disableServices(), for sharing with installPreConfigured.
* In TaskFactory, have tasks optionally declare a "profile" under which
  they will be registered, so that tasks can run under either Installer
  or installPreConfigured or both.

Bug: T352113
Change-Id: I057b0260421567e213bc0a31c55f7c7b6a667d06