Remove User::canExist()
This function has confusing semantics, mixing username validity, user
existence, and whether the user is allowed to log in (isUsable()).

The intended use case: "a central user exists and processing is being
done with the user name, but the orresponding local user on a wiki may
not yet have been created" is better served by the newly introduced
CentralIdLookup::isOwned() in 49f4bab141d0.

The method was added in the 1.43 release cycle in 074be0f04b13, so we
can still remove it without deprecation or release notes.

Depends-On: I800ce1bf1c9f5068168469768065834a8e03747f
Change-Id: I5a36ee4f058476f7dabab0a36c44bc74c8a64d7a