objectcache: Remove `wanobjectcache.$keygroup.regen_set_delay` metric
Introduced in 2019 with I053a73b40d (97e0939082) as part of an
optimization effort around spikes of set operations for the same
keys (T203786). This has since been mitigated in other ways, with
the related cool-off bounce feature removed in 2023 with Ie3a2215d33
(T321634, f83f75d3e6).

As it stands, we have:

* getwithset_seconds{result=hit} which effectively measures the
  typical memcached-get latency.
* getwithset_seconds{result=miss} which effectively measures
  memcached-get + regen callback + memcached-set.
* regen_seconds which effectively measures the regen callback.

Which seems sufficient. We can also eyeball `miss - get - regen` to
gauge the latency of the "set" operation, but we don't have a need
to measure that precisely for every memc call in production, idem
for the "get + regen" subset that regen_set_delay was measuring.

Change-Id: I1c59466f0407f6b980fd04a64a7ab2e3f25c5b0d