rdbms: Make sure ILBFactory::getLocalDomainID() returns a string
Something in T374660 is seemingly causing it to return a non-string
value – I’m guessing it’s an unconfigured PHPUnit mock that defaults to
returning null. Let’s add a static return type so that PHPUnit will know
that it has to return a string, as the phpdoc already claimed. (Note
that the real implementation in LBFactory is guaranteed to return a
string, as DatabaseDomain::getId() has a static return type already.)

Bug: T374660
Change-Id: I0c961343f8f3e83106eee5b47e421c36e36836cf
Depends-On: Ic30aa7d91475b54660153e6cf56556f9cdb1cfd3