Fix PostgreSQL database creation
DatabaseInstaller::needsUpgrade() called selectDatabase() to check for
the existence of a database, explictly rethrowing a connection error.
This does not work on PostgreSQL because selectDatabase() needs to
reconnect and the nonexistence of a database is a connection error. The
whole install process fails due to the uncaught exception.

So, just use getConnection(CONN_CREATE_SCHEMA) and ignore connection
errors. This works for all databases, if we just tweak
SqliteInstaller::openConnection() to not implicitly create a database
file.

Change-Id: I27d39b1567a17582f77edaa0d5bca82edb62418b