mediawiki.special.block [Vue]: fix form submission and add tests
Add a blockType v-model to BlockTypeField, and use that to determine if
we are placing a partial or sitewide block. Also set the default to be
'sitewide'.

Give ExpiryField datetime input the form name `wpExpiry-other` to match
current Special:Block.

Introduce mockMwApiGet() to the test suite which provides mocks for the
known API calls, and warns about those that aren't. Additional mocks or
overrides can be provided per-test as needed. This system hopefully will
help make the Jest suite more robust with realistic data to work with.

Bug: T358934
Change-Id: I5e1baaeb4dc9ae8ef9fcb44513b1c6a675eb94d1