mediawiki.special.block [Vue]: surface server-side msgs + loading state
This adds a new message box for successful messages, and reworks the old
one to handle all error messages.

Error messages are inline for now (like the old Special:Block). Note
these are only the server-side messages, that chiefly serve as a
fallback for clientside error states which will be tackled in a future
patch.

A loading state is added for the block API request by simply disabling
the form fields. Once completed, the user is scrolled to the top to view
the appropriate messaging. This uses "smooth" scrolling so that it's a
bit less "in your face", but this is subject to design review. For now,
it's just important the user is brought to the approriate error message.

Other changes:
* Add a bit of left margin to the inline error message, consistent with
  the no-JS version of this page. Without this, things "jump" a bit as
  the Vue applicaton loads. More refinements for this to come later.
* Fix styles for .mw-block-partial-options to be exactly the same as
  Codex labels.
* Remove unnecessary IIFE from init.js, and add a class to the form
  so we can style it closer to the designs.

Bug: T369471
Bug: T358934
Change-Id: I47344e9581439289642c9424a5f229929805cdbe