maintenance: add transaction round wrapper methods to Maintenance
These help with the basic pattern of running loop of transaction
rounds, each of which should handle deferred updates and possible
replication wait barriers.

Add Maintenance::newBatchIterator() helper method for easily wrapping
an iterator with an batching iterator. This is useful for grouping the
writes in transaction rounds and avoiding the slowness of excessive
waitForReplication() calls.

Update some maintenance scripts to use these methods, including:
* makeTestEdits.php
* importSiteScripts.php
* cleanupSpam.php
* protect.php
* undelete.php
* rollbackEdits.php
* namespaceDupes.php
* uppercaseTitleForUnicodeTransition.php

Deprecate the single-connection transaction wrappers.

Bug: T373057
Change-Id: I42256f9e775bad102e6e5d47c512e0b5cc60d070