Log pages relying on Skin to add Codex styles
This code was added for templates, but it's also applying to
MediaWiki code which is using Html::messageBox without the associated
requirement around adding styling via ResourceLoader.

Apparently various pages are assuming that Codex styles
are loaded by default which is incorrect. The skin should
not be responsible for adding these styles where needed and when it
does it makes the user download duplicate CSS for message box styles
so we should remove this as soon as we can.

If we remove it, we would likely break rendering on various pages and
it's hard to identify code that is using Html::messageBox
and not adding Codex style modules without doing so manually, so this logging should help.

It works on the assumption that all names will likely
feature the string "codex". This should likely help us identify the
majority of cases.

Bug: T375287
Change-Id: I7842600b9ee67a304ee8396842a0c2e2e2519e99