Add Southern Ndebele language (nr)
This is done experimentally, with only a Names.php entry and
namespace translations, before we have actual localized strings.
For details, see https://www.mediawiki.org/wiki/Future_of_Language_Incubation.

I've tested it on my local MediaWiki instance, and it works
as expected: no user interface translations, but the namespaces
appear corretly in page titles, Special:AllPages, etc.,
when `$wgLanguageCode = 'nr'` is set in LocalSettings.php,
and the language name appears correctly in the dropdown in the
"Internationalisation" section in the Preferences.

Bug: T375377
Change-Id: I44cf6a6dabf2ade873506a78573db4f7783a9c99