resourceloader: Serve images with "Access-Control-Allow-Origin: *"
This allows them to be used in 'mask-image' rules in CSS (introduced
in e977eea1533c) even if the image is loaded from a different domain.

It also allows JavaScript code to e.g. draw the images on <canvas>
and read back the pixel data. This should be fine, the images don't
contain any private data.

Bug: T371530
Change-Id: Iec3bdd91ca094e2da5030ac5f16ae96be49bf78e