Fix styling of message boxes on create account page, adds stable API
The stable API mw.util.messageBox is added. This matches the existing
Html::messageBox method and provides a stable way to render messages
on the client going forward. Since these are often used by gadgets
and in extensions this seems like a good idea to allow us to easily
change the HTML in future without breaking things and to detect usage
more easily.

Bug: T375519
Change-Id: If864607cbece899222afba57ca4bfe2f72f9dcfe