Tests: Monitor total bundle size of article views
This new test tracks combined bundle sizes for article views.

It is currently non-blocking and for now prints to the output
because of limitations with CI. We are exploring options for
making this more visible.

This also allows extensions/skins to opt-out of bundle checks
by adding a bundlesize.config.json where modules are declared as
having null bundle size.

Bug: T360590
Change-Id: I67ff5393f8fb5d4e8da58cb2ffbc58207abac2cd