rdbms: Remove reliance on internal STATUS_TRX and STATUS_SESS_ sequence
Follows-up I39cee1363496dfe2, which added a condition based on
`status > ERROR` which I think requires a reliance on, and
understanding/remembering of, the numerical sequence of statuses.

This sequence appears undocumented, both in the TransactionManager
constants and the trxStatus/sessionStatus methods, but either way is not
needed in any of the current cases. This made the code harder to
understand, increases doubt about its correctness, requiring additional
checking while debugging or code reviewing.

Change-Id: Ia0f4b0cfae02de52e7618e558cca6d266689e760