specialpage: Fix incorrect transformation of authentication fields
This patch fixes a bug in the `getFieldDefinitions()` method that
incorrectly applies transformations to the username and/or password
fields when the form fields are coming from authentication request
objects.

When generating form fields from authentication request objects that
have either a username or password, we don't want to apply such a
transformation and mess with redirect based authentication workflow
that injects data via a return request.

We want to be able to pass in a return request in a redirect workflow
using a parameter field that can be loaded on form submission and
receive the data matching the field defined. In the case of 'username',
or 'password', we don't want to apply these transformation if the we
dealing with forms coming from authentication requests.

To test this locally, I added a `username` query param to the return
request and tried seeing if while continuing authentication, I have
the username field set to what is supplied rather than it being `null`.

Bug: T369641
Change-Id: I60e5d44c6651fc4501686130648238947bc7b5a0