Navigation: allow integers in PagerNavigationBuilder queries
The existing `?string` typehint causes phan to complain when integers
are passed. However, these values are only used to build a query string,
and the code works just as fine with integers. On top of that, some of
these things are commonly integers anyway (things like limits and
offsets), so it makes sense to accept integers explicitly.

Change-Id: I5c6b8ad8075d295666a6a04d8c95398dfd9f4060