Mark now unused getDirMarkEntity as deprecated
In HTML context the recommendation is to use <bdi> HTML tag
instead of adding a hidden character that eventually will
end up be in clipboard as described in T27277 also and this
is now unused so let's just mark this as deprecated.

There are some getDirMark() uses however that are in HTML
which should be replaced with appropiate HTML tag which
is unrelated to this deprecation anyway.

Bug: T375975
Change-Id: I5cc15ff00ade5795428c7897df23e4dbac6147a5