ParserTestRunner: consistent normalization of "known failure" output
Improve the output when a parser test fails because the "known failure"
output differs from the actual output, and use the same normalization
function which Parsoid's test runner does for this comparison.

Previously it would display the actual/expected output from the test,
which had no relationship to the "real" reason the test was failing.

Depends-On: I7ff5b27415d98e45d1364161ed6cdaac2d156a81
Change-Id: I0d56b60abb78e37d539267f744afb52c092cb997