Use HTML markup instead of bidi control chars in Special:NewPages
Page link bidi separation is done above use <bdi>, there wasn't a
need for separation between links and length, both are part of the
UI and $ulink is getting a <bdi> itself in Linker::userLink with
<bdi>.

Generally to understand and review these changes keep this rule
of thumb in mind that whenever a user generated content such as
username, page title, summary and content is used in one line
with text and messages that are part of software UI of MediaWiki,
those user generated content should be wrapped with <bdi>. It's
like XSS and SQL Injection but for bidi and the antidote is an
appropiate wrapping with HTML <bdi> tag.

Comment also need a bidi but it already doesn't have it and it
should be applied inside the paranthesis so it will applied
separately.

Bug: T375975
Change-Id: Ic70568c4232a2d283030eb6f60de4e9b6775035e