Mark getDirMark as deprecated
So far I've only seen very limited correct use of getDirMark
outside HTML which one was for the {{DIRMARK}} magic word,
majority of other uses were replaceable with actual HTML markup
so let's use soft deprecation to signal third party developers
they also can go through the same way and use HTML markup
instead of hidden control characters. In future if the magicword
was the only valid use of this we can just hard code this there
and hard deprecate and eventually remove this and if it had more
uses that made sense to keep this around we can undeprecate it.

Bug: T375975
Change-Id: Ie9b78102d58e81fbc6cb22fb85da4962ee2c70b5