objectcache: Add support for sister keys in SqlBagOStuff
This would allow us to force parsercache idhash and idoptions keys next
to each other so when e.g. depooling 1 of 4 hosts in a parser cache cluster,
the amount of cache misses due to rehashing will be ~25% instead of ~%50.

This is similar to the implementation in WANCache and the same structure
has been used to make it consistent.

Bug: T373037
Change-Id: I920fe76e45298aeee6acf725324a5e1ed2b57a37