signup: Fix AJAX username/password check
Fix form error display logic, doubly broken in If864607cbece8992:
it used too much array dereferencing in the error.length === 1 case
so only the first character of the string was shown, and it passed
a HTML string to messageBox() which expects a DOM node or plain
text. (Ideally, there would be some mechanism similar to
OO.ui.HtmlSnippet that allows passing HTML, but that's left for the
future.)

Also deduplicate the error formatting logic.

Bug: T376428
Change-Id: I053b4bd0cf3d6a0962540a90ae7f6b25129362a4