Add static return type for `ParserOutput::getExternalLinks`
PHPUnit tests that mock the ParserOutput object are unable to
correctly infer that the mock should return an empty array rather
than null for `getExternalLinks`. This is currently causing test
failures in SpamBlacklist in CI.

Add the return type definition to the function field definition
so that PHPUnit has a better chance at doing the right thing.

Note that `getExternalLinks` returns `$this->mExternalLinks` by
reference; if there’s some existing code which reassigns a non-array
value to that reference (and, consequently, to `$this->mExternalLinks`,
such code will start to throw TypeErrors during the assignment.

Bug: T376633
Change-Id: I246d5541200c9d0c405f30ea9de091ff9c0e759c