Revert "Use HTML markup instead of bidi control chars in wiki changes"
The change made for T375975 caused an important visual regression in
mobile watchlist, some text is overlapping (T376814). As a result the
MediaWiki train is blocked.

A fix has been proposed I62b64a9503bd3888668503278d4d33d068d58863.
Meanwhile revert the feature from the deployment branch. It will be
deployed next week with its fix if it is merged by that time.

This reverts commit 596c2355aadccbc7a79edee6455bf055f669037f.

Bug: T375975
Bug: T376814
Change-Id: Ic89c353099a548b8627457458bec8622a15204d8