Sort tests in PHPUnit Parallel test suites by name
The parallel test suite group builder (`TestSuiteBuilder`) currently
adds tests to the split groups in the order that the test classes
are encountered in the output of phpunit's `--list-tests-xml`. This
ordering is unspecified. This may lead to the split groups being
composed differently in different environments.

Sort the test classes by filename before adding them to the split
groups so that we have a deterministic ordering and composition of
splits.

Bug: T375851
Change-Id: I6e573028dbcfb19125c40ed2e25259e49fbf59a0