user: Add a warning when creating cross-wiki users
UserIdentity has a wiki property, but User does not implement it
in a meaningful way so there's information loss when converting
from UserIdentityValue to User (and on wiki farms where the same
user can have different names on different wikis, we might even
end up acting on the wrong user). Warn about it.

Adds a new 'user' log channel, which is not ideal, but no existing
channel seemed even remotely relevant.

Change-Id: I895241b8823de9944ca212e538efbfd277bb1f8a