block: Invalidate sessions on BlockDisablesLogin blocks
Use SessionManager::invalidateSessionsForUser() rather then just
resetting the user token when logging out a user who is blocked
on a $wgBlockDisablesLogin wiki, so they are properly logged out
even when an authentication extension like CentralAuth is active.

Doesn't entirely fix the todo in that that it will invalidate
sessions for the local user with the same name as the blocked
user, and if that's a cross-wiki user there's still no guarantee
it is the same user. But in practice it's unlikely not to be.

Change-Id: Ic48814fcfeb71ca0736bac5f133ff407cf494021