ParserCache: Allow for gradual roll out of the new key scheme
This makes sure all entries for the same page end up in the same
database table in the same cluster so depool/crash of a parsercache host
wouldn't have out of proportion effects on the cache overall.

But if we just change the key scheme, every key will be displaced and
everything will go down. So we need to introduce a temporary config variable
to gradually roll out the change.

Bug: T373037
Change-Id: Iae9b8dd5dd65c6d7c8d3b6f786a110d72f0b959e