installer: Make installer initial content be configurable
With the new installPreConfigured.php, it's possible to customise the
installer using configuration. Add a configuration variable allowing the
user to customise the initial page content.

Add {{InstallerOption:}} and {{InstallerConfig:}} fake parser functions,
for feature parity with addWiki.php. Production needs a message like
"This subdomain is reserved for the creation of a $site in $lang". These
variables can be provided to InitialContentTask as installer options.

Bug: T352113
Change-Id: I6f72505cd6d5cb31b4d0105aeea31e1e2f1997de