Use foreach+array_map for iterating over two arrays at the same time
To iterate over the corresponding values from two or more arrays at
the same time, one can use this pattern:

    foreach( array_map( null, $as, $bs ) as [ $a, $b ] ) { ... }

Use it instead of iterating manually with a while loop and the
primitives reset()/current()/next() in a couple of cases.
I find it easier to read, and it's definitely less error-prone.

Change-Id: Iaf1fcf7ed7591b492160e99c2f4ba4be54de2e1a