installer: Move key generation to getDefaultSettings
Key generation just involves setting a couple of variables for later use
in LocalSettingsGenerator. After T374591, install steps will be fairly
heavy. So move this to getDefaultSettings(). A comment indicated that a
warning could be emitted, but this has not been the case for years.

Remove special case in WebInstaller::doUpgrade(), no longer needed.

Remove the generateKeys() call in WebInstallerExistingWiki and tweak the
form so that it doesn't reflect the generated key back to the user.

Bug: T374591
Change-Id: I65670421c51e04719d63eba6c8cd71943c973499