installer: Remove "envPreps" concept
There was a comment justifying the existence of envPreps which wasn't
quite correct. Also the method list abstraction seems like overkill when
nothing is done for each step, like return value checking.

So, fold envPreps into getDefaultSettings().

Add getDefaultSettingsOverrides() as a migration path in case anyone is
using the overrides mechanism to append to envPreps.

Migrate the override of envPrepPath in WebInstaller to its own special
method called by getDefaultSettings().

Change-Id: I093290afaac2d1a343b6e4eca522ffe38ee62587