SpecialBlock: Move additional options to their own component
Move the four additional block options to their own component
(called AdditionalDetailsField), so
we can have easier control over displaying them etc. All the logic
for showing or hiding is moved to the block store.

The existing BlockDetailsOptions component can be changed in a
separate patch to only handle the block details.

Bug: T377509
Change-Id: I2b4c7eaeded68435df04faf1df915572948ffae8