SpecialBlock [Vue]: sync target user selection and make input clearable
Rework the TargetBlockLog component to use the Composition API and the
Pinia store. Update Pinia store on selection changes.

Add the 'clearable' option to the Lookup component, allowing the user to
"reset" things that listen to the target user field, such as the block
log.

Also hide table headings if there are no results.

Bug: T377707
Change-Id: Ia3745ffe858152c6152b1d3b37c6d19d9fb4b712