Sort namespaces list vertically on Special:Search
Currently, the patch makes previous 4-column layout into whatever amount of columns can fit the block. If that is necessary to restore to what it was before, it can be done with CSS adjustment.

Spacing is also improved a bit considering OOUI adds 12px margin-top by default, which is close to 1em (14px) added by Special:Search. This patch increases it to 1.5em for readability.

Bug: T377900
Change-Id: Ic9f4c513bf58e08f4e8eab40f0e2a4a54d1972d8