installer: Reverse disableStorage() by reloading service wiring
MediaWikiServices::disableStorage() works by overriding service wiring
for a number of services. The documentation indicates that
resetGlobalInstance() can be used to reverse this, but that has not been
true since 2016. Installer::resetMediaWikiServices() tries to work
around this by supplying its own wiring for one of the disabled
services, but this is fragile.

So, add a reload option to resetGlobalInstance() which reloads the
service wiring, and use it in the installer. Rename the $quick parameter
to $mode.

Bug: T374591
Change-Id: I9777760d019952d9a92af0e6b68865bdaa83545f