maintenance: make waitForReplication() respect any injected ILBFactory
Rename setDBProvider() method to setLBFactory() and make it require an
ILBFactory (which implements IConnectionProvider). Add a getLBFactory()
method and make getReplicaDB(), getPrimaryDB(), and waitForReplication()
call it. This way, they all use the same LBFactory.

Note that setDBProvider() had no callers and was marked internal. In any
case, replacing setDB() with setLBFactory() will be useful in the future.
It is rare enough for scripts to inject DB connections that working with
ILBFactory is not too difficult.

Change-Id: I6f4df240c15703c66c835e723f762635a77180af