Remove getText usages in the vicinity of MessageCache
This is the fifth patch of a series of patches to remove
ParserOutput::getText() calls from core. This series of patches should
be functionally equivalent to I2b4bcddb234f10fd8592570cb0496adf3271328e.

Here we handle the case of MessageCache, which requires a bit of
refactoring to start deprecating the fact that MessageCache::parse
can return either a string (if not a top-level parse) or a ParserOutput
(at top-level), which callers typically use to decide whether to run
getText on the output or not. We provide a
MessageCache::parseWithPostprocessing method that handles this
case and runs the pipeline at the desired level, and returns a
ParserOutput always.

Bug: T293512
Change-Id: I8a1fa84e650d920d07d74722d8059d5afeedec6b