specials: Do not pass boolean to Message::numParams
Password policy settings 'value' can be a number, boolean or a null

But only number needs to be passed to the corresponding policy
messages (to be output). Boolean and null represent the message
assertion and are not output in the eventual message string (the
messages do not have parameter).

Values with false and null are already skipped, so this change only
affects values with boolean true.

Bug: T378133
Change-Id: Ib9e6f7ed14f9982f9d2350ccd7b212f0fbf10477