Remove remaining uses of getText in core
This is the sixth patch of a series of patches to remove
ParserOutput::getText() calls from core. This series of patches should
be functionally equivalent to I2b4bcddb234f10fd8592570cb0496adf3271328e.

This patch adds an access to ParserOptions to HtmlOutputRendererHelper
so that it can be passed by the caller and re-used after the rendering,
for instance to post-process the generated ParserOutput. This is used
in compareLanguageConverterOutput.

Bug: T293512
Change-Id: I6e58430cdad571268c21ef97e407e32a570777a1