objectcache: Remove docs about RedisConnectionPool configurable size
There exists no option matching this description.

The class was created in Feb 2013 with Ie72536240e (03833b54e8),
which initially included a `poolSize` option that matches this
description.

This was removed the next month in March 2013,
with I6ea53378c9 (646051d36d).

While at it, fix file headers and ingroup doc annotations

> Add missing `ingroup` to class blocks (and remove from file blocks)
> as otherwise the file is indexed twice (e.g. in Doxygen) which makes
> navigation on doc.wikimedia.org rather messy.
>
> Remove duplicate descriptions from file blocks in favour of class
> doc blocks. This reduces needless duplication and avoids incorrect
> or outdated copies emerging over time, and helps make file headers
> more consistently (visually) ignorable.
>
> Ref https://gerrit.wikimedia.org/r/q/message:ingroup+is:merged+owner:Krinkle+branch:master

Change-Id: I285122054c8848d0fa700aa323ff813d5f0c4690