Category: Refresh counts for negative page count
The current code checks for negative member count (mPages contains the
sum of subcats, files and pages), checking also for negative page count
and adjust the member count if there is a wrong value.
Set the member count after correction of subcats and files count to
avoid in worse case another stale number.

Bug: T373773
Change-Id: Ia8d6bdf6f68152396544b8a8dd73c3922839a9f9