Support 'include' parameter in namespace select inputs
Namespace select inputs already support an 'exclude' parameter, ignoring
any namespaces that match ids in that array. For
Special:GlobalContributions, it would be useful to have the opposite -
an array of namespaces to be included, as it only wants to support
filtering on common namespaces.

- Allow an 'include' parameter to be passed through to the namespace
  select input
- Support 'include' in Html inputs and OOUI inputs

TODO: Support 'include in Codex inputs (see T378582)

Bug: T378179
Change-Id: I4170beb588292fbec26926ca3afe559c25de9f55