debug: Remove use of makeConfigSetScript() from MWDebug
There is no need to change the minification status of this blob
based on ResourceLoader debug mode as it is rendered visually
in the debug toolbar (wgDebugToolbar) already.

Removing this removes one of the last remaining indirect calls
to global static ResourceLoader::inDebugMode().

Alternatively, we could unconditionally JSON_PRETTY_PRINT this,
since it naturally limited to debug mode. I choose not to because
in 15 years I don't recall ever needing to debug the contents of this
blob. By always pretty printing it, it would further dominate
view-source when debugging anything else in MediaWiki locally. If
truly needed, it is trivial to pass pretty=true, or var_dump, or
inspect via PhpStorm, any part of this when actually working on this.

Bug: T36738
Change-Id: Id2ccd63e20c11fb63652262eb7e9994e5d15ae74