Allocate tests to groups sequentially instead of round-robin
The round-robin allocation of tests to `split_groups` resulted in
`split_groups` containing tests from a large number of different
extensions, and the composition of `split_groups` changing
significantly depending on which extensions are currently loaded.

Instead, we now add tests to buckets in their alphabetical order so
that tests are more or less grouped by extension. At the same time
we try to preserve the time balance of the buckets.

Bug: T378478
Change-Id: I1da2848b041a2b6a879c64f3a3ee8dd9471fd06e