SpecialBlock: Add suppress log accordion
Show suppress log entries in an accordion below the 'Recent blocks'
accordion. This list is not limited to only 'block' and 'reblock'
because the view-more link already doesn't; this may need further
work.

Empty-state messages are added for the new suppress log as well
as for active logs (which was using the same message as previous
blocks).

Bug: T377725
Change-Id: I63ae67a9bd384a9a128a356e1be4a250c6114b2d