Add ParserOutputFlags::{HAS_ASYNC_CONTENT,ASYNC_NOT_READY}
This allows Parsoid to mark parses which contain async content which
is "not ready yet".  At the moment this output is cached with a reduced
TTL, although in the future it might still be treated as uncacheable,
cached until evicted, or some other option.

The HAS_ASYNC_CONTENT flag along with ParserOutput::hasReducedExpiry()
ensures that RefreshLinksJob is opportunistically reinvoked whenever
the page is reparsed, since the asynchronous content may change the
metadata for the page when it becomes ready.

As describe in T373256, ::hasReducedExpiry() is misnamed now, and a
follow-up patch will probably rename it to ::hasDynamicContent() or
something like that.  What it really means is "RefreshLinksJob must
be re-run on every parse, because the content may change on each
parse".  In the past we would *also* reduce the cache time for
pages like this.  But for asynchronous content, "the content may
change on each parse" only *until* the asynchronous content is
"ready".  Once it is ready the contents will no longer change, and
the cache lifetime can be raised again -- but ::hasDynamicContent()
still needs to be set, which in the future will mean "you need to
check that RefreshLinksJob has last run" not "you must always run
RefreshLinksJob".

Asynchronous content will always set HAS_ASYNC_CONTENT, even after
the content is "ready", but will only set ASYNC_NOT_READY if it
needed to use placeholder content in this render.

Bug: T373256
Change-Id: I71e10f8a9133c16ebd9120c23c965b9ff20dabd2